backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytube==15.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
git+https://github.com/openai/whisper.git
python-dotenv==1.0.0
openai>=1.6.1
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.18
youtube_transcript_api==0.6.2
ffmpeg-python